requirements.txt
# 📌 Note: PyAudio may fail on Windows if installed via pip.
# ✅ To install PyAudio, use:
# pip install pipwin
# pipwin install pyaudio

absl-py==2.3.1
aiofiles==24.1.0
annotated-types==0.7.0
anyio==3.7.1
astunparse==1.6.3
audioread==3.0.1
beautifulsoup4==4.13.4
birdnetlib==0.13.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.7.9
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
docopt==0.6.2
fastapi==0.104.1
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
flatbuffers==25.2.10
fonttools==4.58.5
fsspec==2025.5.1
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.176.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
greenlet==3.2.3
grpcio==1.73.1
grpcio-status==1.71.2
h11==0.16.0
h5py==3.14.0
hf-xet==1.1.5
httpcore==1.0.9
httplib2==0.22.0
httpx==0.28.1
huggingface-hub==0.33.4
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
joblib==1.5.1
Js2Py==0.74
keras==2.15.0
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.1
llvmlite==0.44.0
Markdown==3.8.2
MarkupSafe==3.0.2
matplotlib==3.10.3
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
numba==0.61.2
numpy==1.26.4
oauthlib==3.3.1
openai==1.95.1
opencv-python==4.8.1.78
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.1
pillow==11.3.0
pipwin==0.5.2
platformdirs==4.3.8
pooch==1.8.2
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pygame==2.6.1
pyjsparser==2.7.1
pyparsing==3.2.3
PyPrind==2.11.3
pySmartDL==1.3.4
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-oauthlib==2.0.0
resampy==0.4.2
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
six==1.17.0
sniffio==1.3.1
soundfile==0.12.1
soupsieve==2.7
soxr==0.5.0.post1
SQLAlchemy==2.0.41
starlette==0.27.0
sympy==1.14.0
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
watchdog==2.1.9
websockets==12.0
Werkzeug==3.1.3
wrapt==1.14.1
python-dotenv==1.0.1
